client channel aborts on unknown error from service dispatch
when a client channel attempts to dispatch to a service and gets an unexpected return
it will now return an error

it was retrying which, when the service crashed on input, was just killing all instances of the service.
this change is based on the assumption that bad input crashing a service will likely crash all services, so abort the bad input.