new pubsub replacement (WARN: completely new client facing pubsub api)
!!! The client api for pubsub has changed, please see README.md for
!!! the new details of how to use pubsub.

The old pub_sub had a number of problems:

+ topics had to be added in the config file
+ routed all messages through a single node: adding latency and allowing a
  temporary single point of failure (only lasting as long as it takes for
  all publisher and subscribers to open new channels)

The new approach fixes both of these. There's now a 'subs' p6dmap table
that for each topic has all the subscribers of that topic. The new
'pub' service dumps out all the subscribers for that topic and fires
a message to each one of them.

This does mean the publishing mmd might be pushing more data then it would
have in the past, although should be minor compared to all pubsub traffic
going through an unlucky node as we had in the past.