sub.chans service
This service provides notification whenever a subscriber starts or
stops subscribing on a particular topic, much like how services
notifies when services are registered and unregistered.

This could be used for publishers that might have to do expensive
processing to publish, and only want to bother if there's a
subscriber. It could also be used for implimenting a client-side dmap
where you want to be able to kill any data published by a node if/when
it goes away.