Added modify checks + server flags module
Now honors If-Modified-Since and will send 304 (not modified) if the browser timestamp isn't identical to the server's.
will resend if the file is older or newer.

added mmd_web_flags, (hit URL http://localhost:9998/mmd_flags/)
uses a cookie to set flags to change the behaivor of file vending, logging
etc